feat: Study 시작종료 관련 모델 및 레포지토리 구현

+++ core/remote/src/main/java/com/benenfeldt/remote/dto/StudyStartResponse.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class StudyStartResponse(
-    val id: Int,
+    val id: Long,
     val startTime: String,
-    val userId: Int,
+    val userId: Long,
 )
